requirements.txt
hazm @ https://github.com/roshan-research/hazm/archive/master.zip
wordcloud==1.9.2
arabic-reshaper==3.0.0
python-bidi==0.4.2